+++ app/Http/Controllers/SessaoTreinamentoController.php
@@ -10,12 +10,16 @@ use App\Models\SerieExercicio;
 use App\Models\SessaoTreinamento;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class SessaoTreinamentoController extends Controller
 {
 
     public function index()
     {
+        if (!Gate::allows('viewAny')) {
+            abort(403, 'Acesso não autorizado.');
+        }
         // Buscar todas as sessões de treinamento,banco de dados
         $programasTreinamento = ProgramaTreinamento::all();
         $sessoesTreinamento = SessaoTreinamento::all();
@@ -24,20 +28,20 @@ class SessaoTreinamentoController extends Controller
 
     public function create()
     {
+        if (!Gate::allows('create')) {
+            abort(403, 'Acesso não autorizado.');
+        }
         $programasTreinamento = ProgramaTreinamento::all();
         return view('sessao_treinamento.sessao_treinamento_create',compact('programasTreinamento'));
     }
 
     public function store(SessaoTreinamentoRequest $request)
     {
-        ///dd($request);
-        ///  dd($request);
-        //DB::enableQueryLog(); // Habilita o log de queries
-        // Os dados já foram validados e preparados
+        if (!Gate::allows('create')) {
+            abort(403, 'Acesso não autorizado.');
+        }
         $dadosValidados = $request->validated();
-        // Lógica para salvar os dados no banco
 
-        // SessaoTreinamento::create($dadosValidados);
         try {
             SessaoTreinamento::Create($dadosValidados);
             // Sucesso
@@ -50,30 +54,30 @@ class SessaoTreinamentoController extends Controller
 
     public function show(SessaoTreinamento $sessao)
     {
+        if (!Gate::allows('view')) {
+            abort(403, 'Acesso não autorizado.');
+        }
         return view('sessoes.show', compact('sessao'));
     }
 
     public function edit($idSessaoTreinamento)
     {
+        if (!Gate::allows('update')) {
+            abort(403, 'Acesso não autorizado.');
+        }
 //        $sessaoTreinamento = SessaoTreinamento::findOrFail($idSessaoTreinamento);
         $sessaoTreinamento = SessaoTreinamento::with('programaTreinamento')->findOrFail($idSessaoTreinamento);
 //        dd($sessaoTreinamento);
-
         $programaTreinamento = $sessaoTreinamento->programaTreinamento;
-//        dd('Teste');
-//        dd($programaTreinamento);
         return view('sessao_treinamento.sessao_treinamento_edite', compact('sessaoTreinamento','programaTreinamento'));
     }
 
     public function update(SessaoTreinamentoRequest  $request, $idSessaoTreinamento)
     {
-      ///  dd($request);
-        //DB::enableQueryLog(); // Habilita o log de queries
-        // Os dados já foram validados e preparados
+        if (!Gate::allows('update')) {
+            abort(403, 'Acesso não autorizado.');
+        }
         $dadosValidados = $request->validated();
-        // Lógica para salvar os dados no banco
-      //  dd($dadosValidados);
-       // SessaoTreinamento::create($dadosValidados);
         try {
             SessaoTreinamento::updateOrCreate(
                 ['id_sessao_treinamento' => $idSessaoTreinamento],
@@ -89,6 +93,9 @@ class SessaoTreinamentoController extends Controller
 
     public function destroy($idSessaoTreinamento)
     {
+        if (!Gate::allows('delete')) {
+            abort(403, 'Acesso não autorizado.');
+        }
         SessaoTreinamento::destroy($idSessaoTreinamento);
         return redirect()->route('sessao_treinamento.index')->with('success', 'Sessão de treinamento deletada com sucesso.');
     }
